requirements.txt
gradio==3.0
transformers==4.11.3
torch==2.0.0
openai-whisper==20230918